+++ HISTORY.rst
@@ -8,6 +8,8 @@ Pending
 
 * New release notes here
 * Now MySQL 5.7 compatible
+* The final message from ``SmartChunkedIterator`` is now rounded to the nearest
+  second.
 
 1.0.5 (2016-02-10)
 ------------------